Implement ability to search all fields.
Implement the ability to search all fields from the document.

Todo: Bugfix

Current code is correctly finding entries by search term. (As
confirmed by System.out statements. (For instance it finds four
matches for PHP). However, the print statement is printing the
last line of the CSV file equal to the matches. Fix code so that
it is printing the result returned by findByValue.